+++ xqsight-manage/xqsight-manage-service/src/main/java/com/xqsight/sys/service/SysRoleService.java
@@ -1,12 +1,9 @@
-/**
- * 上海汽车集团财务有限责任公司
- * Copyright (c) 1994-2015 All Rights Reserved.
- */
 package com.xqsight.sys.service;
+
+import com.xqsight.sys.model.SysRole;
 
 import java.util.List;
 
-import com.xqsight.sys.model.SysRole;
 
 /**
  * @Description: TODO
@@ -47,7 +44,7 @@ public interface SysRoleService {
 	 * @return void    返回类型
 	 * @throws
 	 */
-	void deleteSysRole(int roleId);
+	void deleteSysRole(long roleId);
 	
 	/**
 	 * 
@@ -73,6 +70,6 @@ public interface SysRoleService {
 	 * @return SysRole    返回类型
 	 * @throws
 	 */
-	SysRole querySysRoleByRoleId(int roleId);
+	SysRole querySysRoleByRoleId(long roleId);
 
 }
