+++ xqsight-manage/xqsight-manage-service/src/main/java/com/xqsight/sys/service/impl/SysRoleServiceImpl.java
@@ -1,21 +1,16 @@
-/**
- * 上海汽车集团财务有限责任公司
- * Copyright (c) 1994-2015 All Rights Reserved.
- */
 package com.xqsight.sys.service.impl;
 
-import java.util.List;
-
+import com.alibaba.fastjson.JSON;
+import com.xqsight.sys.model.SysRole;
+import com.xqsight.sys.mysqlmapper.SysRoleMapper;
+import com.xqsight.sys.service.SysRoleService;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.alibaba.fastjson.JSON;
-import com.xqsight.sys.model.SysRole;
-import com.xqsight.sys.mysqlmapper.SysRoleMapper;
-import com.xqsight.sys.service.SysRoleService;
+import java.util.List;
 
 /**
  * @Description: TODO
@@ -34,7 +29,6 @@ public class SysRoleServiceImpl implements SysRoleService {
 	 * <p>Title: saveSysRole</p>
 	 * <p>Description: </p>
 	 * @param sysRole
-	 * @see com.xqsight.sys.service.SysRoleService#saveSysRole(SysRole)
 	 */
 	@Override
 	public void saveSysRole(SysRole sysRole) {
@@ -46,7 +40,6 @@ public class SysRoleServiceImpl implements SysRoleService {
 	 * <p>Title: updateSysRole</p>
 	 * <p>Description: </p>
 	 * @param sysRole
-	 * @see com.xqsight.sys.service.SysRoleService#updateSysRole(SysRole)
 	 */
 	@Override
 	public void updateSysRole(SysRole sysRole) {
@@ -58,10 +51,9 @@ public class SysRoleServiceImpl implements SysRoleService {
 	 * <p>Title: deleteSysRole</p>
 	 * <p>Description: </p>
 	 * @param roleId
-	 * @see com.xqsight.sys.service.SysRoleService#deleteSysRole(int)
 	 */
 	@Override
-	public void deleteSysRole(int roleId) {
+	public void deleteSysRole(long roleId) {
 		logger.debug("出入参数:{}", roleId);
 		sysRoleMapper.deleteSysRole(roleId);
 	}
@@ -71,7 +63,6 @@ public class SysRoleServiceImpl implements SysRoleService {
 	 * <p>Description: </p>
 	 * @param roleName
 	 * @return
-	 * @see com.xqsight.sys.service.SysRoleService#querySysRoleByRoleName(String)
 	 */
 	@Override
 	public List<SysRole> querySysRoleByRoleName(String roleName) {
@@ -90,10 +81,9 @@ public class SysRoleServiceImpl implements SysRoleService {
 	 * <p>Description: </p>
 	 * @param roleId
 	 * @return
-	 * @see com.xqsight.sys.service.SysRoleService#querySysRoleByRoleId(int)
 	 */
 	@Override
-	public SysRole querySysRoleByRoleId(int roleId) {
+	public SysRole querySysRoleByRoleId(long roleId) {
 		return sysRoleMapper.querySysRoleByRoleId(roleId);
 	}
